[MIG] website_whatsapp: Migration to 16.0
