+++ website_whatsapp/views/res_config_settings.xml
@@ -4,8 +4,11 @@
         <field name="model">res.config.settings</field>
         <field name="inherit_id" ref="website.res_config_settings_view_form" />
         <field name="arch" type="xml">
-            <xpath expr="//div[@id='social_media_settings']" position="after">
-                <div class="col-12 col-lg-6 o_setting_box" id="website_setting">
+            <xpath expr="//div[@id='website_email_marketing']" position="inside">
+                <div
+                    class="col-12 col-lg-6 o_setting_box"
+                    id="website_whatsapp_setting"
+                >
                     <div class="o_setting_left_pane">
                         <field name="whatsapp_enabled" />
                     </div>
